Improve directions for "Name" field of LP snaps
I recently had a user turn up very confused because the docs website
said (or at least strongly implied) that this field has to be the name
of the snap; but they were trying to create two snap recipes for
different git branches that uploaded to different tracks in the store,
in which case a blanket "use the same name as the name of the snap"
instruction doesn't work.

These are getting-started instructions, so I don't want to overload them
with detail that many people won't need, but I've tweaked the wording to
try to indicate that the name of the snap is a good default but you can
use something else if need be.